+++ spec/dynamic_spec.rb
@@ -44,6 +44,51 @@ describe ELFTools::Dynamic do
     end
   end
 
+  describe 'relocations' do
+    def elf(name)
+      ELFTools::ELFFile.new(File.open(File.join(__dir__, 'files', name)))
+    end
+
+    def from_sections(file)
+      file.sections.select { |sec| sec.is_a?(ELFTools::Sections::RelocationSection) }.flat_map(&:relocations)
+    end
+
+    def summarize(relocations)
+      relocations.map { |rel| [rel.header.r_offset.to_i, rel.symbol_index, rel.type_name] }
+    end
+
+    it 'reads what the sections record' do
+      # Both views describe the same bytes, so each answers for the other.
+      %w[amd64.elf i386.elf aarch64.elf ppc64.elf libc.so.6].each do |name|
+        file = elf(name)
+        expect(summarize(file.dynamic.relocations)).to eq summarize(from_sections(file))
+        expect(file.dynamic.relocations).not_to be_empty
+      end
+    end
+
+    it 'reads them from a file that has no sections left' do
+      data = File.binread(File.join(__dir__, 'files', 'amd64.elf'))
+      header = ELFTools::ELFFile.new(StringIO.new(data)).header
+      header.e_shoff = 0
+      header.e_shnum = 0
+      header.e_shstrndx = 0
+      data[0, header.num_bytes] = header.to_binary_s
+      file = ELFTools::ELFFile.new(StringIO.new(data))
+      expect(file.sections).to be_empty
+      expect(summarize(file.dynamic.relocations)).to eq summarize(from_sections(elf('amd64.elf')))
+    end
+
+    it 'reports a table that is not loaded' do
+      data = File.binread(File.join(__dir__, 'files', 'amd64.elf'))
+      tag = ELFTools::ELFFile.new(StringIO.new(data)).dynamic.tag_by_type(:jmprel)
+      tag.header.d_val = 0xdeadbeef000
+      data[tag.header.offset, tag.header.num_bytes] = tag.header.to_binary_s
+      file = ELFTools::ELFFile.new(StringIO.new(data))
+      expect { file.dynamic.relocations }
+        .to raise_error(ELFTools::ELFError, 'Invalid DT_JMPREL address 0xdeadbeef000')
+    end
+  end
+
   describe 'broken DT_STRTAB' do
     # Returns an ELF file whose DT_STRTAB tag has been modified by +block+.
     # @yieldparam [ELFTools::Structs::ELF_Dyn] header The DT_STRTAB header.